server/requirements.txt
PyJWT==2.8.0
python-dotenv==1.0.0
bcrypt==4.1.2
sanic==23.6.0
Sanic_Cors==2.2.0
asyncpg==0.29.0
sqlalchemy==2.0.28